backend/services/response_service.py
```python
from typing import List, Optional
from backend.database.response_repository import ResponseRepository
from backend.models.response_model import Response
from backend.services.analysis_service import AnalysisService
from backend.utils.auth_utils import AuthUtils

class ResponseService:
    def __init__(self):
        self.response_repository = ResponseRepository()
        self.analysis_service = AnalysisService()

    def submit_response(self, response_data: dict, user_id: Optional[int] = None) -> Response:
        # Create a new response object
        response = Response(**response_data)
        if user_id:
            response.user_id = user_id

        # Save the response to the database
        saved_response = self.response_repository.add_response(response)

        # Return the saved response object
        return saved_response

    def get_responses_by_survey_id(self, survey_id: int) -> List[Response]:
        # Retrieve responses for a specific survey
        responses = self.response_repository.get_responses_by_survey_id(survey_id)
        return responses

    def analyze_survey_responses(self, survey_id: int):
        # Retrieve responses for the survey
        responses = self.get_responses_by_survey_id(survey_id)

        # Analyze the responses using the AnalysisService
        analysis_results = self.analysis_service.analyze_responses(responses)

        # Return the analysis results
        return analysis_results

    def delete_response(self, response_id: int) -> bool:
        # Delete a response by its ID
        return self.response_repository.delete_response(response_id)

    def update_response(self, response_id: int, update_data: dict) -> Response:
        # Update a response with new data
        updated_response = self.response_repository.update_response(response_id, update_data)
        return updated_response

    def get_response_by_id(self, response_id: int) -> Response:
        # Retrieve a single response by its ID
        response = self.response_repository.get_response_by_id(response_id)
        return response

    def get_responses_for_user(self, user_id: int) -> List[Response]:
        # Retrieve all responses submitted by a specific user
        responses = self.response_repository.get_responses_for_user(user_id)
        return responses

    def is_response_owner(self, response_id: int, user_id: int) -> bool:
        # Check if the user is the owner of the response
        response = self.get_response_by_id(response_id)
        return response.user_id == user_id

    def validate_response_access(self, response_id: int, token: str) -> bool:
        # Validate if the user has access to the response
        user_id = AuthUtils.decode_auth_token(token)
        return self.is_response_owner(response_id, user_id)
```